fixes completions for the current module
Because we currently filter according to the imports in the current
module, we don't get any completions for identifiers defined in the
current module, because a module doesn't import itself.

This fixes this issue while also using the new psc-ide-get-module-name
function to make a proper guess on which module to load with the load
command.