Determine whether the server is running by making a conn check
This is the most (only?) reliable way to judge whether the server is
running at any given point in time, and if it's cheap enough to start
a network connection for every individual request (e.g. for
flycheck/eldoc) then it's cheap enough to ping the server briefly as
needed.

The new function `psc-ide-test-connection` benchmarks at ~1ms on my
machine, so even though it might be called fairly frequently (on
modeline update, for example) then this should be negligible.

In the event that the overhead ends up being excessive, we can use a
var to cache the last successful connection time, and check that var
first. Note that this is tricky to do correctly in future if we plan
to support separate connections on a per-project basis.